feat: 문자열 덧셈 계산기 뼈대 코드 작성
