+++ src/main/java/calculator/SplitStrategy.java
@@ -1,0 +1,5 @@
+package calculator;
+
+public interface SplitStrategy {
+    String[] split(String input);
+}
